Change start-up config to place URIs
Previously, the start-up configuration reads keys of the pattern
'store-X-dir'. With the introduction of an in-memory store, these are
specified using special URIs. And 'store' is sometimes a misleading
name. At least for the configuration this term was renamed to 'place'.
Consequently, the start-up config keys follows now the pattern
'place-X-uri'.